submissions/jayan/requirements.txt
# Data & utils
pandas>=2.1.0
numpy>=1.25.0
joblib>=1.3.0
matplotlib>=3.8.0

# ML
scikit-learn>=1.3.0
xgboost>=1.8.2

# Transformers / LLM
transformers>=4.30.0
sentencepiece>=0.1.98
torch>=2.1.0
accelerate>=0.26.0

# SHAP + visualization
shap>=0.41.0
pillow>=10.0.0

# Web UI
gradio>=3.39.0
